Move models in db to models folder
