+++ db/models/Project.js
@@ -1,9 +1,9 @@
 const Joi = require('@hapi/joi');
 const { DataTypes, Model } = require('sequelize');
-const sequelize = require('../config/sequelize');
+const sequelize = require('../../config/sequelize');
 const Tag = require('./Tag');
 const Link = require('./Link');
-const { removeImg } = require('../utils/fileSystem');
+const { removeImg } = require('../../utils/fileSystem');
 
 const STRICT_PROJECT_SCHEMA = Joi.object({
   id: Joi.string().uuid(),
